+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Added `updateData` function
 - Added `refreshData` function
+- Added `refresh` option
+- Added `stopRefresh` function
 - Remove colors from tooltips for Chart.js pie chart
 
 ## 2.1.2